requirements.txt
django
pillow
djangorestframework
psycopg2